program work fine

+++ DataBase Manager/gui/mainwindow_wi.py
@@ -11,14 +11,15 @@ from typing import Generator, Iterable
 
 # IMPORT CORE
 from PySide6.QtCore import QTimer
-from PySide6.QtWidgets import QFileDialog, QTableWidgetItem
+from PySide6.QtWidgets import QFileDialog, QTableWidgetItem, QMessageBox
 
 # IMPORT UI
 try:
     from gui.__dbmanager import Ui_MainWindow as _mainWindow
     from PySide6.QtWidgets import QMainWindow as _base
 
-except ImportError:
+except ImportError as err:
+    print(f"{err =}\t{type(err).__name__}")
 
     from PySide6.QtUiTools import loadUiType
     _path = iPath.dirname(__file__)
@@ -33,16 +34,22 @@ locate = iPath.dirname(__file__)
 # MAIN WINDOW CLASS
 class MainWindow(_base, _mainWindow):
     """
+    DataBase Manager MainWindow
     """
+    VERSION: tuple[str, str, str] = ('1.00-R', 'Ready', ':)')
+
     CURRENT_PATH = locate
     FILE_PATH:str = None
     FORMAT: str = None
 
     SQL_TABLE_LOAD: str = None
+    SQL_LOAD_COLUMN: str = None
+    ALL_TABLE: list[str] = None
     CONN: sqlite3.Connection = None
 
     def __init__(self, *args, **kwargs) -> None:
         """
+        Initialize DataBase Manager Simple Gui
         """
         # Initialize
         super(MainWindow, self).__init__(*args, **kwargs)
@@ -57,6 +64,7 @@ class MainWindow(_base, _mainWindow):
 
     def __actions(self) -> None:
         """
+        Connect Actions Signal
         """
         # File Menu
         self.actionNew.triggered.connect(self.__new)
@@ -64,32 +72,117 @@ class MainWindow(_base, _mainWindow):
         self.actionSave.triggered.connect(self.__save)
         self.actionSaveAs.triggered.connect(self.__saveAs)
         self.actionExit.triggered.connect(self.__exit)
+      
         # Table Menu
+        self.actionNewTable.triggered.connect(self.__newTable)
+        self.actionLoadTable.triggered.connect(self.loadSqlite)
+        self.actionSaveTable.triggered.connect(self.__save)
         self.actionAddRow.triggered.connect(self.__addRow)
         self.actionAddColumn.triggered.connect(self.__addColumn)
         self.actionDelRow.triggered.connect(self.__delRow)
         self.actionDelColumn.triggered.connect(self.__delColumn)
+      
         # Help Menu
         self.actionAbout.triggered.connect(self.__about)
 
         # Signal Connect
         self.cmb_tableName.currentTextChanged.connect(self.__tableChange)
 
-    def __new(self) -> None:
+        # Button Connect
+        self.btn_newTable.clicked.connect(self.__newTable)
+        self.btn_loadTable.clicked.connect(self.loadSqlite)
+        self.btn_saveTable.clicked.connect(self.__save)
+
+    def __algorithm(self) -> None:
         """
+        Set Algorithm Name And Set 'Enabled'
         """
-        self.__saveAs()
+        _name: dict = {'csv': "CSV", 'sqlite': "SQLITE-3", None: 'None'}
+        self.lbl_algorithm.setText(_name[self.FORMAT])
         
         if self.FORMAT == 'csv':
-            with open(self.FILE_PATH, 'w') as f:
-                f.close()
+           
+            self.btn_loadTable.setEnabled(False)
+            self.btn_saveTable.setEnabled(False)
+            self.btn_newTable.setEnabled(False)
+            self.cmb_tableName.setEnabled(False)
+
+            self.actionNewTable.setEnabled(False)
+            self.actionSaveTable.setEnabled(False)
+            self.actionLoadTable.setEnabled(False)
+
+            self.actionAddColumn.setEnabled(True)
+            self.actionDelColumn.setEnabled(True)
+
+        elif self.FORMAT == 'sqlite':
+           
+            self.btn_loadTable.setEnabled(True)
+            self.btn_saveTable.setEnabled(True)
+            self.btn_newTable.setEnabled(True)
+            self.cmb_tableName.setEnabled(True)
+
+            self.actionNewTable.setEnabled(True)
+            self.actionSaveTable.setEnabled(True)
+            self.actionLoadTable.setEnabled(True)
+
+            self.actionAddColumn.setEnabled(False)
+            self.actionDelColumn.setEnabled(False)
+
+    def __newTable(self) -> None:
+        """
+        Clear Table for New One
+        """
+        self.tableWidget.clear()
+        self.tableWidget.setColumnCount(10)
+        self.tableWidget.setRowCount(10)
+
+    def __new(self) -> None:
+        """
+        New File
+        """
+        
+        if self.FILE_PATH != None:
+            saveQuestion = QMessageBox(self)
+            act = saveQuestion.question(self, "Save Question", "Do You Want to Save Table ?", QMessageBox.Yes, QMessageBox.No)
+
+            if act == saveQuestion.Yes:
+                self.__save()
+            else:
+                saveQuestion.close()
+
+        self.FILE_PATH = None
+        self.ALL_TABLE = None
+        self.FORMAT = None
+        self.CONN = None
+        self.__algorithm()
+        self.cmb_tableName.clear()
+        self.tableWidget.clear()
+        self.tableWidget.setColumnCount(10)
+        self.tableWidget.setRowCount(10)
+
+        self.__saveAs()
+
+        if self.FORMAT == 'csv':
+            if iPath.isfile(self.FILE_PATH) != True and self.FILE_PATH != '':
+                with open(self.FILE_PATH, 'w') as f:
+                    f.close()
         
         if self.FORMAT == 'sqlite':
             self.db_sqlite.connect(self.FILE_PATH)
 
     def __open(self) -> None:
         """
+        Open Dialog
         """
+        if self.FILE_PATH != None:
+            saveQuestion = QMessageBox(self)
+            act = saveQuestion.question(self, "Save Question", "Do You Want to Save Table ?", QMessageBox.Yes, QMessageBox.No)
+
+            if act == saveQuestion.Yes:
+                self.__save()
+            else:
+                saveQuestion.close()
+
         filter: str = "Sqlite3 (*.db);;Csv (*.csv);;Custom Sqlite3 (*.*);;Custom Table (*.*)"
         path, form = self.dialogFiles.getOpenFileName(self, "Load Table", self.CURRENT_PATH, filter)
         self.FILE_PATH = path
@@ -97,23 +190,32 @@ class MainWindow(_base, _mainWindow):
         
         #print(f"{path= }\t{form = }")
         if path != '':
+            self.cmb_tableName.clear()
             if self.FORMAT == 'sqlite':
                 self.addTableName()
             
             elif self.FORMAT == 'csv':
                 self.loadCsv()
+            
+            self.__algorithm()
 
     def __save(self) -> None:
         """
+        Save Table
         """
-        if self.FORMAT == 'sqlite':
-            self.exportSqlite(self.FILE_PATH)
+        if self.FILE_PATH != None:
+            if self.FORMAT == 'sqlite':
+                self.exportSqlite(self.FILE_PATH)
         
-        elif self.FORMAT == 'csv':
-            self.exportCsv(self.FILE_PATH)
+            elif self.FORMAT == 'csv':
+                self.exportCsv(self.FILE_PATH)
+        
+        else:
+            self.__saveAs()
 
     def __saveAs(self) -> None:
         """
+        Save As Other Path Table
         """
         filter: str = "Sqlite3 (*.db);;Csv (*.csv);;Custom Sqlite3 (*.*);;Custom Table (*.*)"
         path, form = self.dialogFiles.getSaveFileName(self, "Save Table", self.CURRENT_PATH, filter)
@@ -127,6 +229,8 @@ class MainWindow(_base, _mainWindow):
             
             elif self.FORMAT == 'csv':
                 self.exportCsv(path)
+            
+            self.__algorithm()
 
     def __addRow(self) -> None:
         """
@@ -150,7 +254,6 @@ class MainWindow(_base, _mainWindow):
         """
         self.tableWidget.removeColumn(self.tableWidget.currentColumn())
 
-
     def __about(self) -> None:
         """
         """
@@ -163,18 +266,20 @@ class MainWindow(_base, _mainWindow):
 
     def __tableChange(self, table: str) -> None:
         """
+        Sqlite Table Change
         """
-        self.SQL_TABLE_LOAD = table
-       # print(f"{self.SQL_TABLE_LOAD= }")
-        self.loadSqlite()
+        if self.FORMAT == 'sqlite':
+            self.SQL_TABLE_LOAD = table
 
     def toTable(self, data) -> None:
         """
         Insert Data To Table
         """
-        self.tableWidget.clear()
+        if self.FORMAT == 'csv':
+            self.tableWidget.clear()
+        
         self.tableWidget.setRowCount(0)
-        #print("--> TO TABLE <--")
+
         sizeLines: list[int] = []
         for row, item in enumerate(data):
             sizeLines.append(len(item))
@@ -191,25 +296,31 @@ class MainWindow(_base, _mainWindow):
         """
         Add Table Name To ComboBox
         """
+        self.cmb_tableName.clear()
+        self.cmb_tableName.addItem('')
+
         self.CONN = self.db_sqlite.connect(self.FILE_PATH)
         get_tableName = self.CONN.execute("SELECT name FROM sqlite_master WHERE type = ? AND name NOT LIKE ?", ('table', 'sqlite_%'))
-        tableName: list[str] = [i[0] for i in get_tableName]
+        self.ALL_TABLE: list[str] = [i[0] for i in get_tableName]
+
+        script_get_pattern: str = "SELECT * FROM sqlite_master WHERE type = 'table' AND name NOT LIKE 'sqlite_%';"
+        self.SQL_LOAD_COLUMN: str = tuple(self.CONN.execute(script_get_pattern))[0][-1]
+        print(f"{self.SQL_LOAD_COLUMN= }")
         
-        #print(f"{tableName= }")
-        
-        for t in tableName:
+        for t in self.ALL_TABLE:
             self.cmb_tableName.addItem(t)
 
     def existsTable(self) -> bool:
         """
         If Exists Table 'True' Else 'False'
         """
-        if self.SQL_TABLE_LOAD is not None:
-            script: str = f"SELECT count(*) FROM sqlite_master WHERE type='table' AND name='{self.SQL_TABLE_LOAD}'"
-            exists = bool(next(self.CONN.execute(script))[0])
-            #print(f"{exists= }")
+        if self.ALL_TABLE is not None:
+            if self.SQL_TABLE_LOAD in self.ALL_TABLE:
+                script: str = f"SELECT count(*) FROM sqlite_master WHERE type='table' AND name='{self.SQL_TABLE_LOAD}'"
+                exists = bool(next(self.CONN.execute(script))[0])
+                #print(f"{exists= }")
             
-            return exists
+                return exists
         
         return False
 
@@ -219,6 +330,7 @@ class MainWindow(_base, _mainWindow):
         """
         script = f"SELECT name FROM PRAGMA_TABLE_INFO('{self.SQL_TABLE_LOAD}');"
         getLabel: list = [l[0] for l in self.CONN.execute(script)]
+        self.tableWidget.clear()
         self.tableWidget.setColumnCount(len(getLabel))
         self.tableWidget.setHorizontalHeaderLabels(getLabel)
         
@@ -243,6 +355,14 @@ class MainWindow(_base, _mainWindow):
         """
         Load Data From Sqlite3 DataBase
         """
+        if self.SQL_TABLE_LOAD != None:
+            saveQuestion = QMessageBox(self)
+            act = saveQuestion.question(self, "Save Question", "Do You Want to Save Table ?", QMessageBox.Yes, QMessageBox.No)
+
+            if act == saveQuestion.Yes:
+                self.__save()
+            else:
+                saveQuestion.close()
         
         if self.existsTable():
             self.setHeaders()
@@ -289,6 +409,7 @@ class MainWindow(_base, _mainWindow):
 
     def getFromTable(self) -> Generator:
         """
+        Get Data From Table
         """
         column_range: int = self.tableWidget.columnCount()
         row_range: int = self.tableWidget.rowCount()
@@ -308,16 +429,51 @@ class MainWindow(_base, _mainWindow):
 
     def updateSqliteTable(self) -> None:
         """
+        Update Table Value
         """
-        pass
+        getColumn: tuple = tuple(self.tableWidget.horizontalHeaderItem(i).text() for i in range(0, self.tableWidget.columnCount()))
+        #print(f"{getColumn= }")
+        
+        valS: str = ", ".join((f':{num}' for num in range(0, len(getColumn))))
+        rows: Generator = self.getFromTable()
+        
+        toMap: dict = lambda x: {f"{i}": v for i, v in enumerate(x)}
+
+        for row in rows:
+            update_script: str = f"INSERT OR REPLACE INTO {self.SQL_TABLE_LOAD} VALUES({valS})"
+            self.CONN.execute(update_script, toMap(row))
+            self.CONN.commit()
+            
+            #print(f"updateRow >> {row= }")
 
     def addSqliteTable(self) -> None:
         """
+        Create New Table In DataBase And Set Data
         """
-        pass
+        rows: Generator = self.getFromTable()
+        
+        column: tuple[str] = tuple(next(rows)) if self.SQL_LOAD_COLUMN is None else None
+        
+        script_create: str = f"CREATE TABLE IF NOT EXISTS {self.SQL_TABLE_LOAD}{column}" if self.SQL_LOAD_COLUMN is None else self.SQL_LOAD_COLUMN
+        
+        self.CONN.execute(script_create)
+        
+        columnValid: tuple = tuple(self.tableWidget.horizontalHeaderItem(i).text() for i in range(0, self.tableWidget.columnCount())) if column is None else tuple(column)
+        valS: str = ", ".join(tuple(f':{num}' for num in range(0, len(columnValid))))
+        script_insert: str = f"INSERT OR REPLACE INTO {self.SQL_TABLE_LOAD} VALUES ({valS})"
+        
+       # print(f"{script_insert=}")
+
+        toMap: dict = lambda x: {f"{i}": v for i, v in enumerate(x)}
+
+        for row in rows:
+           # print(f"ROW > {num= }")
+            self.CONN.execute(script_insert, toMap(row))
+            self.CONN.commit()
 
     def exportSqlite(self, file_path: str) -> None:
         """
+        Export Data To Sqlite DataBase to 'File_Path' Only Load Table Saved
         """
         if iPath.exists(file_path):
             if self.existsTable():
@@ -343,6 +499,17 @@ class MainWindow(_base, _mainWindow):
         """
         Export Table To CSV File or Other File Table to 'File_Path'
         """
+        get_header: list[str] = [self.tableWidget.horizontalHeaderItem(i) for i in range(0, self.tableWidget.columnCount())]
+        valid: str = lambda x: x.text() if x is not None else '1'
+        all_decimal: Iterable[bool] = [valid(header).isdecimal() for header in get_header]
+        use_header: bool = False if False not in all_decimal else True
+
+        del all_decimal, valid
+
+        if use_header:
+            get_header: tuple = tuple(it.text() for it in get_header)
+            header_valid: str = "{}\n".format('\t'.join(get_header))
+
         getItem: Generator = (
             '\t'.join(row).rstrip()
             for row in self.getFromTable()
@@ -350,6 +517,8 @@ class MainWindow(_base, _mainWindow):
 
         try:
             with open(file_path, 'w') as f:
+                if use_header:
+                    f.write(header_valid)
                 f.write('\n'.join(getItem).rstrip())
         
         except (FileNotFoundError, FileExistsError, IsADirectoryError, IOError) as errfile:
@@ -357,5 +526,5 @@ class MainWindow(_base, _mainWindow):
         
         finally:
             f.close()
-            del getItem, f
+            del getItem, f, get_header, header_valid, use_header
